Create README.md
9 lines (47 sloc)  3.77 KB
0x05 Processes and signals 🔧
Using a shell script is most useful for repetitive tasks that may be time consuming to execute by typing one line at a time. A few examples of applications shell scripts can be used for include: Automating the code compiling process. Running a program or creating a program environment. This project covers processes and signals

At the end of this project, I was able to solve these questions:

What is a PID
What is a process
How to find a process’ PID
How to kill a process
What is a signal
What are the 2 signals that cannot be ignored